+++ src/main/java/com/github/pagehelper/PageInterceptor.java
@@ -117,7 +117,7 @@ public class PageInterceptor implements Interceptor {
     public void setProperties(Properties properties) {
         //缓存 count ms
         msCountMap = CacheFactory.createSqlCache(properties.getProperty("msCountCache"), "ms", properties);
-        String dialectClass = properties.getProperty("dialect");
+        String dialectClass = properties.getProperty("impl");
         if (StringUtil.isEmpty(dialectClass)) {
             dialectClass = default_dialect_class;
         }
