+++ src/main/java/com/github/pagehelper/dialect/helper/impl/HsqldbDialect.java
@@ -1,31 +1,29 @@
-package com.github.pagehelper.helper.dialect;
+package com.github.pagehelper.dialect.helper.impl;
 
 import com.github.pagehelper.Page;
-import com.github.pagehelper.helper.HelperDialect;
+import com.github.pagehelper.dialect.helper.HelperDialect;
 import org.apache.ibatis.cache.CacheKey;
 import org.apache.ibatis.session.RowBounds;
 
 /**
  * @author liuzh
  */
-public class MySqlDialect extends HelperDialect {
+public class HsqldbDialect extends HelperDialect {
 
     @Override
     public String getPageSql(String sql, Page page, RowBounds rowBounds, CacheKey pageKey) {
-        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 14);
+        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 20);
         sqlBuilder.append(sql);
-        if (page.getStartRow() == 0) {
+        if (page.getPageSize() > 0) {
             sqlBuilder.append(" LIMIT ");
             sqlBuilder.append(page.getPageSize());
-        } else {
-            sqlBuilder.append(" LIMIT ");
+            pageKey.update(page.getPageSize());
+        }
+        if (page.getStartRow() > 0) {
+            sqlBuilder.append(" OFFSET ");
             sqlBuilder.append(page.getStartRow());
-            sqlBuilder.append(",");
-            sqlBuilder.append(page.getPageSize());
             pageKey.update(page.getStartRow());
         }
-        pageKey.update(page.getPageSize());
         return sqlBuilder.toString();
     }
-
 }
