+++ src/main/java/com/github/pagehelper/dialect/helper/impl/MySqlDialect.java
@@ -1,30 +1,30 @@
-package com.github.pagehelper.helper.dialect;
+package com.github.pagehelper.dialect.helper.impl;
 
 import com.github.pagehelper.Page;
+import com.github.pagehelper.dialect.helper.HelperDialect;
 import org.apache.ibatis.cache.CacheKey;
 import org.apache.ibatis.session.RowBounds;
 
 /**
  * @author liuzh
  */
-public class SqlServer2012Dialect extends SqlServerDialect {
+public class MySqlDialect extends HelperDialect {
 
     @Override
     public String getPageSql(String sql, Page page, RowBounds rowBounds, CacheKey pageKey) {
         StringBuilder sqlBuilder = new StringBuilder(sql.length() + 14);
         sqlBuilder.append(sql);
-        if (page.getStartRow() > 0) {
-            sqlBuilder.append(" OFFSET ");
+        if (page.getStartRow() == 0) {
+            sqlBuilder.append(" LIMIT ");
+            sqlBuilder.append(page.getPageSize());
+        } else {
+            sqlBuilder.append(" LIMIT ");
             sqlBuilder.append(page.getStartRow());
-            sqlBuilder.append(" ROWS ");
+            sqlBuilder.append(",");
+            sqlBuilder.append(page.getPageSize());
             pageKey.update(page.getStartRow());
         }
-        if (page.getPageSize() > 0) {
-            sqlBuilder.append(" FETCH NEXT ");
-            sqlBuilder.append(page.getPageSize());
-            sqlBuilder.append(" ROWS ONLY");
-            pageKey.update(page.getPageSize());
-        }
+        pageKey.update(page.getPageSize());
         return sqlBuilder.toString();
     }
